// src/app/interfaces/user-registration.interface.ts
import { User } from "../model/user.model";
import { Creneau } from "./creaneau.interface";
import { Espace } from "./espace.interface";
import { Poste } from "./poste.interface";

export interface UserRegistration {
    id: number;
    isAffected: boolean;
    isAccepted: boolean;
    user: User;
    Espace: Espace;
    Creneau: Creneau;
  }